# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Financial Management API Documentation",
    "version": "1.0.0",
    "description": "This is the API documentation for the Financial Management application, providing endpoints for managing accounts, transactions, and budgets."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/transactions": {
      "get": {
        "summary": "Retrieve all transactions",
        "description": "Get a list of all financial transactions from the database",
        "tags": [
          "Transactions"
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transactions retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Transaction"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new transaction",
        "tags": [
          "Transactions"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Transaction"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/transactions/{id}": {
      "get": {
        "summary": "Get transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Transaction ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction retrieved successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Transaction"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Update transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Transaction ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete transaction by ID",
        "tags": [
          "Transactions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Transaction ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Transaction deleted successfully"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "required": [
          "userId",
          "accountId",
          "amount",
          "description",
          "date",
          "category",
          "type",
          "currency"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the transaction"
          },
          "userId": {
            "type": "string",
            "description": "ID of the user who owns the transaction"
          },
          "accountId": {
            "type": "string",
            "description": "ID of the account associated with the transaction"
          },
          "amount": {
            "type": "number",
            "description": "Transaction amount"
          },
          "description": {
            "type": "string",
            "description": "Description of the transaction"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date of the transaction"
          },
          "category": {
            "type": "string",
            "description": "Category of the transaction"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "description": "Type of transaction"
          },
          "currency": {
            "type": "string",
            "description": "Currency code (e.g., USD, EUR)"
          },
          "convertedAmount": {
            "type": "number",
            "description": "Amount converted to base currency"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "example": {
          "id": "txn_123",
          "userId": "user_456",
          "accountId": "acc_789",
          "amount": 100.5,
          "description": "Grocery shopping",
          "date": "2023-10-01T10:00:00.000Z",
          "category": "Food",
          "type": "expense",
          "currency": "USD",
          "convertedAmount": 100.5,
          "createdAt": "2023-10-01T10:00:00.000Z",
          "updatedAt": "2023-10-01T10:00:00.000Z"
        }
      },
      "CreateTransactionRequest": {
        "type": "object",
        "required": [
          "accountId",
          "amount",
          "description",
          "date",
          "category",
          "type",
          "currency"
        ],
        "properties": {
          "accountId": {
            "type": "string",
            "description": "ID of the account for the transaction"
          },
          "amount": {
            "type": "number",
            "description": "Transaction amount"
          },
          "description": {
            "type": "string",
            "description": "Description of the transaction"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date of the transaction"
          },
          "category": {
            "type": "string",
            "description": "Category of the transaction"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "description": "Type of transaction"
          },
          "currency": {
            "type": "string",
            "description": "Currency code"
          }
        },
        "example": {
          "accountId": "acc_789",
          "amount": 100.5,
          "description": "Grocery shopping",
          "date": "2023-10-01T10:00:00.000Z",
          "category": "Food",
          "type": "expense",
          "currency": "USD"
        }
      },
      "UpdateTransactionRequest": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "ID of the account for the transaction"
          },
          "amount": {
            "type": "number",
            "description": "Transaction amount"
          },
          "description": {
            "type": "string",
            "description": "Description of the transaction"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "description": "Date of the transaction"
          },
          "category": {
            "type": "string",
            "description": "Category of the transaction"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "description": "Type of transaction"
          },
          "currency": {
            "type": "string",
            "description": "Currency code"
          }
        },
        "example": {
          "amount": 150,
          "description": "Updated grocery shopping",
          "category": "Groceries"
        }
      }
    }
  },
...
}